feat!: Rename frontmatter publish properties
We want to unabbreviate "pub" in our frontmatter, so we were wondering
what it actually stood for;

    pubDate -> publishDate

could incorrectly suggest that a date in the future delays the post
being published, which is incorrect.

    pubDate -> publishedDate

is accurate but unweieldy to say.

    pubDate -> publishedOn

is both accurate and easy to say, at the cost of not being quite as
similar.

To match this, we can rename publish to published...

    publish -> published

...which also feels somewhat like the NixOS construct

    {
        services.publish.enabled = true;
    }

in its past-tense declarative language (rather than the
more-imperative-sounding "publish"). We also need to rename all publish
-> published and pubDate -> publishedOn in tests to avoid everything
breaking