fix: move <main> to Global
The main element has some useful global styles such as setting the text
and background colors, so moving this to the global file makes sense. We
might not need justification, flexbox, etc. but I consider it simpler to
do the change wholesale than to split up the styles into another div
here.

Change-Id: Ia3238df23422f2fa4e7d2a26823ce33d31b91b49
Reviewed-on: https://review.gerrithub.io/c/FreshlyBakedCake/Whisk/+/1206483
Tested-by: Skyler Grey <skyler3665@gmail.com>
Reviewed-by: Samuel Shuert <me@thecoded.prof>