feat: Improve article formatting
Previously, articles were simply text. For actual use, we probably want
to include other information such as the title and the publish date, as
well as adding some (extremely basic) styles to pad it all properly.

I think we still have missed styling here, for example I suspect there
is not any styling for headings (but they are not used in any of our
test posts, or any private test material)