feat: Improve article formatting
Previously, articles were simply text. For actual use, we probably want
to include other information such as the title and the publish date, as
well as adding some (extremely basic) styles to pad it all properly.

I think we still have missed styling here, for example I suspect there
is not any styling for headings (but they are not used in any of our
test posts, or any private test material)

Change-Id: Ia4ddbc0bbd86dba274eb769736259227153e5711
Reviewed-on: https://review.gerrithub.io/c/FreshlyBakedCake/Whisk/+/1206484
Tested-by: Skyler Grey <skyler3665@gmail.com>
Reviewed-by: Samuel Shuert <me@thecoded.prof>