feat(post): Add link back to main page
If we're viewing a post, we might want a link back to the main
page. This is a bit cleaner than forcing people to cut the last part of
the URL ;)

For extra points on long posts, we can make there also be a link at the
bottom. Another way to solve this would be a navbar (and a similar posts
thing?), and maybe we'll do that in future. Maybe as well. Not for now,
though. Not for now.

Change-Id: I9e0981f4412829672706894b4b5fb3dfdfb198f3