refactor: simplify filepath splitting
I was looking at this statement, and it looked weird to me.

For starters, I thought "you can remove the ternary here", because there
will *always* be one element to this list even if there isn't a slash.
Splitting something without a slash just returns the original element,
so this is equivalent.

    options.entry.split("/").at(-1)

I also considered this library function, which neatly wraps up the same
thing:

    path.basename(options.entry);

But it's from the [NodeJS Library](1), and Astro suggests you shouldn't
use these because [different adapters won't support it](2). Therefore,
I'm using my middle marginally-less-elegant solution, and putting this
"there is a reason for not breaking down Chesterton's fence" into the
commit log.

[1]: https://nodejs.org/api/path.html#pathbasenamepath-suffix
[2]: https://docs.astro.build/en/guides/imports/#node-builtins
Change-Id: Ie2ab1f7245bcecf1c5c603049d14b48a1c97b0eb