feat(main): Improve card design
Previously we had a rather janky placeholder card, it's nicer to have
a list and we can show more information without it looking strange.
Because we no longer use the old card component, I've removed it. There
is, however, a case to be made for factoring out some items (card? tags?
here in future)

Change-Id: If5dca40b41207767a22d3c372227631cd9d053a2